re #1089 reducing build warnings to manageable levels on Windows

+++ trunk/PlusLib/src/Utilities/xio/MahonyAhrsAlgo.cxx
@@ -40,7 +40,7 @@ void MahonyAhrsAlgo::Update(float gx, float gy, float gz, float ax, float ay, fl
     ay *= recipNorm;
     az *= recipNorm;     
 
-    // Normalise magnetometer measurement
+    // Normalize magnetometer measurement
     recipNorm = InvSqrt(mx * mx + my * my + mz * mz);
     mx *= recipNorm;
     my *= recipNorm;
@@ -61,7 +61,7 @@ void MahonyAhrsAlgo::Update(float gx, float gy, float gz, float ax, float ay, fl
         // Reference direction of Earth's magnetic field
         hx = 2.0f * (mx * (0.5f - q2q2 - q3q3) + my * (q1q2 - q0q3) + mz * (q1q3 + q0q2));
         hy = 2.0f * (mx * (q1q2 + q0q3) + my * (0.5f - q1q1 - q3q3) + mz * (q2q3 - q0q1));
-        bx = sqrt(hx * hx + hy * hy);
+        bx = sqrtf(hx * hx + hy * hy);
         bz = 2.0f * (mx * (q1q3 - q0q2) + my * (q2q3 + q0q1) + mz * (0.5f - q1q1 - q2q2));
 
     // Estimated direction of gravity and magnetic field
